Add jdbc connection handling code
`ActiveRecordSqlServerAdapter::Jdbc` contains all code that is needed
for the jdbc mode. It does connection and result handling.
Jdbc result values are converted to the corresponding ruby types.

With these changes only the sqljdbc4-java gem has to be installed or
the sqljdbc42.jar has to be included in the java CLASSPATH.